GPU scheduling is updated republish bug is fixed

+++ autoware.ai/src/autoware/common/rubis_sched/include/rubis_sched/sched_profiling.hpp
@@ -22,9 +22,9 @@ namespace sched {
 
   extern struct timespec task_start_time_;
   extern struct timespec task_end_time_;
-  extern unsigned long long gpu_seg_response_time_;
-  extern unsigned long long gpu_seg_execution_time_;
-  extern unsigned long long cpu_seg_response_time_;
+  extern unsigned long gpu_seg_response_time_;
+  extern unsigned long gpu_seg_execution_time_;
+  extern unsigned long cpu_seg_response_time_;
   extern int is_gpu_profiling_ready_;
 
   void init_task_profiling(std::string task_reponse_time_filename);
@@ -36,7 +36,7 @@ namespace sched {
   void start_profiling_gpu_seg_response_time();
   void start_profiling_gpu_seg_execution_time();
   void stop_profiling_gpu_seg_time(unsigned int gpu_seg_id, unsigned int iter, std::string remark = " ");
-  unsigned long long get_current_time_ns();
+  unsigned long get_current_time_ns();
   void start_job_profiling();
   void finish_job_profiling(unsigned int cpu_seg_id);
   void start_gpu_profiling();
