+++ autoware.ai/src/autoware/common/rubis_sched/src/sched_profiling.cpp
@@ -13,9 +13,9 @@ namespace sched{
 
   struct timespec task_start_time_;
   struct timespec task_end_time_;
-  unsigned long long gpu_seg_response_time_;
-  unsigned long long gpu_seg_execution_time_;
-  unsigned long long cpu_seg_response_time_;
+  unsigned long gpu_seg_response_time_;
+  unsigned long gpu_seg_execution_time_;
+  unsigned long cpu_seg_response_time_;
   int is_gpu_profiling_ready_ = 0;
 
   void start_gpu_profiling(){
@@ -91,7 +91,7 @@ namespace sched{
 
   void stop_profiling_gpu_seg_time(unsigned int gpu_seg_id, unsigned int iter, std::string remark){
     if(gpu_profiling_flag_){
-      unsigned long long current_time = get_current_time_ns();
+      unsigned long current_time = get_current_time_ns();
       gpu_seg_response_time_ = current_time - gpu_seg_response_time_;
       gpu_seg_execution_time_ = current_time - gpu_seg_execution_time_;
       
@@ -100,9 +100,9 @@ namespace sched{
     }
   }
 
-  unsigned long long get_current_time_ns(){
+  unsigned long get_current_time_ns(){
     struct timespec ts;
-    unsigned long long current_time;
+    unsigned long current_time;
     clock_gettime(CLOCK_REALTIME, &ts);
     current_time = ts.tv_sec%10000 * 1000000000 + ts.tv_nsec;
     return current_time;
